refactor: move naked-builtin scanning into MissingBuiltinWrapper
The naked_builtin_issues cluster (~95 lines, 5 functions) lived in
Orchestration but was used by exactly one check and produced Credo.Issue
structs - check-layer logic, not shared plumbing. Move it into its sole
consumer, MissingBuiltinWrapper, as private functions. The check is now the
emitter, so the threaded `check` param drops out in favor of the in-check
format_issue/2, and naked_builtin_calls/4 uses a `for` comprehension
instead of flat_map-as-filter. Orchestration shrinks to the genuinely
shared flat_map_*/compiled_check_*/issue helpers.
