+++ lib/ash_credo/check/warning/missing_builtin_wrapper.ex
@@ -44,6 +44,7 @@ defmodule AshCredo.Check.Warning.MissingBuiltinWrapper do
       """
     ]
 
+  alias AshCredo.Introspection
   alias AshCredo.Orchestration
 
   @change_builtins ~w(
@@ -152,7 +153,110 @@ defmodule AshCredo.Check.Warning.MissingBuiltinWrapper do
   @impl true
   def run(%SourceFile{} = source_file, params) do
     Enum.flat_map(@families, fn family ->
-      Orchestration.naked_builtin_issues(source_file, params, __MODULE__, family)
+      naked_builtin_issues(source_file, params, family)
     end)
+  end
+
+  # Flags bare calls to configured builtin functions at statement position
+  # inside action bodies, `pipeline` bodies, and global sections, suggesting
+  # the matching DSL wrapper keyword. Ash's builtins are plain functions
+  # imported into the DSL scope that return spec tuples, so an unwrapped call
+  # is silently discarded and nothing warns at compile time or runtime.
+  #
+  # Required `opts`:
+  #
+  #   * `:builtins` - builtin function names to match
+  #   * `:wrapper` - the DSL keyword to suggest (e.g. `"change"`)
+  #   * `:action_types` - action entity types to scan; scoped to the actions
+  #     whose scope actually imports their builtin family, so a bare call
+  #     elsewhere is a compile error the compiler already catches
+  #
+  # Optional `opts`:
+  #
+  #   * `:pipeline_builtins` - builtin names to also flag at statement position
+  #     inside `pipeline` bodies (defaults to `[]`). Pipeline bodies import all
+  #     three builtin families at once, so families split ownership of names
+  #     that exist in more than one family to avoid double-flagging.
+  #   * `:global_sections` - top-level DSL sections (e.g. `:validations`) whose
+  #     direct entries are scanned for bare builtin calls (defaults to `[]`).
+  #   * `:usage_prefix` - what to put before `builtin(...)` in the advice
+  #     (defaults to `"<wrapper> "`).
+  defp naked_builtin_issues(source_file, params, opts) do
+    builtins = opts |> Keyword.fetch!(:builtins) |> MapSet.new()
+    wrapper = Keyword.fetch!(opts, :wrapper)
+    action_types = Keyword.fetch!(opts, :action_types)
+    pipeline_builtins = opts |> Keyword.get(:pipeline_builtins, []) |> MapSet.new()
+    global_sections = Keyword.get(opts, :global_sections, [])
+    advice = {wrapper, Keyword.get(opts, :usage_prefix, "#{wrapper} ")}
+
+    Orchestration.flat_map_resource_context(source_file, params, fn context, issue_meta ->
+      action_issues =
+        naked_builtin_section_issues(
+          context,
+          {:actions, action_types, builtins},
+          advice,
+          issue_meta
+        )
+
+      pipeline_issues =
+        naked_builtin_section_issues(
+          context,
+          {:pipelines, [:pipeline], pipeline_builtins},
+          advice,
+          issue_meta
+        )
+
+      global_issues =
+        Enum.flat_map(
+          global_sections,
+          &naked_global_section_issues(context, &1, builtins, advice, issue_meta)
+        )
+
+      action_issues ++ pipeline_issues ++ global_issues
+    end)
+  end
+
+  defp naked_global_section_issues(context, section_name, builtins, advice, issue_meta) do
+    case Introspection.resource_section(context, section_name) do
+      nil -> []
+      section_ast -> naked_builtin_calls(section_ast, builtins, advice, issue_meta)
+    end
+  end
+
+  defp naked_builtin_section_issues(
+         context,
+         {section_name, entity_names, builtins},
+         advice,
+         issue_meta
+       ) do
+    with true <- MapSet.size(builtins) > 0,
+         section_ast when not is_nil(section_ast) <-
+           Introspection.resource_section(context, section_name) do
+      section_ast
+      |> Introspection.action_entities(entity_names)
+      |> Enum.flat_map(&naked_builtin_calls(&1, builtins, advice, issue_meta))
+    else
+      _ -> []
+    end
+  end
+
+  # Works on any AST node with a do-block whose direct entries are statements:
+  # action entities, pipeline entities, and global sections.
+  defp naked_builtin_calls(ast, builtins, advice, issue_meta) do
+    for {func_name, meta, _} <- Introspection.entity_body(ast),
+        MapSet.member?(builtins, func_name) do
+      naked_builtin_issue(func_name, advice, meta, issue_meta)
+    end
+  end
+
+  defp naked_builtin_issue(func_name, {wrapper, usage_prefix}, meta, issue_meta) do
+    format_issue(
+      issue_meta,
+      message:
+        "`#{func_name}` has no effect without a `#{wrapper}` wrapper. " <>
+          "Use `#{usage_prefix}#{func_name}(...)` instead.",
+      trigger: "#{func_name}",
+      line_no: meta[:line]
+    )
   end
 end
